+++ apps/client/src/app/lazy-data/+state/lazy-data.selectors.ts
@@ -1,13 +1,13 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { LAZY_DATA_FEATURE_KEY, LazyDataPartialState, State } from './lazy-data.reducer';
+import { LAZY_DATA_FEATURE_KEY, State } from './lazy-data.reducer';
+import { LazyDataKey } from '../lazy-data-types';
 
 // Lookup the 'LazyData' feature state managed by NgRx
-export const getLazyDataState = createFeatureSelector<
-  State>(LAZY_DATA_FEATURE_KEY);
+export const getLazyDataState = createFeatureSelector<State>(LAZY_DATA_FEATURE_KEY);
 
-export const getEntry = createSelector(
+export const getEntry = (key: LazyDataKey) => createSelector(
   getLazyDataState,
-  (state, { key }) => state.data[key]
+  (state) => state.data[key]
 );
 
 export const getEntryRow = createSelector(