+++ apps/client/src/app/pages/log-tracker/fsh-spear-log-order.ts
@@ -1,22 +1,93 @@
-export const fshSpearLogOrder = [
-  'The Lochs',
-  'Loch Seld Northwestern Lakebed',
-  'Loch Seld Central Lakebed',
-  'Loch Seld Southeastern Lakebed',
-  'The Ruby Sea',
-  'Ruby Price Depths',
-  'Tamamizu',
-  'Sui-no-Sato',
-  'The Adventure',
-  'Shisui of the Violet Tides',
-  'The Kobayashi Maru',
-  'Yanxia',
-  'The One River Southwestern Riverbeds',
-  'The One River Southern Riverbeds',
-  'Imperial Hypersonic Assault Craft L-XXIII',
-  'The Sunken Junk',
-  'The Dragon\'s Struggle',
-  'The Azim Steppe',
-  'Azim Khaat Western Lakebed',
-  'Azim Khaat Eastern Lakebed'
+export const fshSpearLogAreas = [
+  {
+    areaName: 'The Lochs',
+    spotNames: [
+      'Loch Seld Northwestern Lakebed',
+      'Loch Seld Central Lakebed',
+      'Loch Seld Southeastern Lakebed',
+      'Loch Seld Deep Lakebed',
+      'Loch Seld Northeastern Lakebed',
+      'Southwestern Loch Seld',
+    ],
+  },
+  {
+    areaName: 'The Ruby Sea',
+    spotNames: [
+      'Ruby Price Depths',
+      'Tamamizu',
+      'Sui-no-Sato', // hyphen '\u002D' (not in db)
+      'Sui–no–Sato', // en dash '\u2013' (in db)
+      'The Adventure',
+      'Shisui of the Violet Tides',
+      'The Kobayashi Maru',
+      'The Eastern Ruby Sea',
+      'Kobayashi Maru\'s Northern Wake',
+      'Northern Onokoro',
+      'Isari Slope',
+      'Exile',
+      'Northeastern Bekko',
+      'Eastern Barrier Isle',
+      'Northwestern Onokoro',
+      'The Anemone Gardens',
+    ],
+  },
+  {
+    areaName: 'Yanxia',
+    spotNames: [
+      'The One River Southwestern Riverbeds',
+      'The One River Southern Riverbeds',
+      'Imperial Hypersonic Assault Craft L-XXIII',
+      'The Sunken Junk',
+      'The Dragon\'s Struggle',
+      'The Sunken Junk\'s South Side',
+    ],
+  },
+  {
+    areaName: 'The Azim Steppe',
+    spotNames: [
+      'Azim Khaat Western Lakebed',
+      'Azim Khaat Eastern Lakebed',
+      'Azim Khaat Southern Lakebed',
+    ],
+  },
+  {
+    areaName: 'Lakeland',
+    spotNames: [
+      'Northeast Source',
+      'The Isle of Ken',
+      'Southeast Source',
+      'Eastern Isle of Ken',
+    ],
+  },
+  {
+    areaName: 'Il Mheg',
+    spotNames: [
+      'North Lyhe Ghiah',
+      'Deepwood Swim',
+      'Central Longmirror Lake',
+      'Thysm Lran',
+      'South Longmirror Lake',
+      'East Longmirror Lake',
+    ],
+  },
+  {
+    areaName: 'The Rak\'tika Greatwood',
+    spotNames: [
+      'North Lake Tusi Mek\'ta',
+      'The Covered Halls of Dwatl',
+      'Central Lake Tusi Mek\'ta',
+      'South Lake Tusi Mek\'ta',
+      'Deep Lake Tusi Mek\'ta',
+    ],
+  },
+  {
+    areaName: 'Thavnair',
+    spotNames: [
+      'Southern Akyaali',
+      'Western Akyaali',
+      'Southern Great Runoff',
+      'The Font of Maya',
+      'Outer Akyaali',
+    ],
+  },
 ];